Revert "Refactor dashboard components: remove MatchesPage and SkillsPage, imp…"
